feat: add avatarUrl to profile fields and move my profile component inside settings
