+++ src/main/java/com/backend/fitters/profile/ProfileRepository.java
@@ -11,7 +11,7 @@ public interface ProfileRepository extends JpaRepository<Profile, Long> {
     @Query(value = """
             SELECT p.zip_code AS zipCode, p.country,
              p.address, p.pricing, p.site, p.bio,
-            p.state, p.specialities
+            p.state, p.specialities, p.avatar_url as avatarUrl
              FROM profile p
              WHERE p.id = :profileId
             """, nativeQuery = true)
