+++ src/main/java/com/backend/fitters/profile/dto/ProfileFieldsDto.java
@@ -20,4 +20,6 @@ public interface ProfileFieldsDto {
 
     String getSpecialities();
 
+    String getAvatarUrl();
+
 }
